when showing the readme, list the files in the sidebar
Automatically showing a readme and files are problematic
because there is no guarantee about where the 2nd of the two will show up.

Solve this by showing the files in the sideabar.
This solution can make it hard to see README headers.
However, a README should generally link out to other documents
and not contain lots of headers itself.

It also creates an incosistency that this is not done for other files.
I think we could make this more consistent in the future
by making file listings and headers collapsible.
If files are shown at the bot