Toggle revision history by clicking on last modified
This is dissapointing since we only have metadata.
There is no way to link to a revision or show a diff.
The only thing that can be offered is an HTML download.

Doing a direct fetch of the HTML download runs into auth issues.
Perhaps there is a way to fix this, but I don't think it is very useful
to show old revisions without any diff.

There is an unofficial URL for getting a diff,
but it is a JSON output that needs rendering.
https://docs.google.com/document/d/{file.id}/showrevision?id={file.id}&end={rev.id}&start={rev.id}