upgrade comments to show threads and users
Comments are now grouped together in threads.
User icons are shown next to the comment.

The styling could certainly be improved.
But I think it as good as what existed previously.

The comments are being replaced one-by-one in the DOM with the newer React version.
It might be preferable to delete all the comments from the DOM
and not do any individual react renderings.
But this mode works well in case there is a problem (such as more than 100 comment threads).
Additionally there will be a future update to how comments are displayed
so that document positioning is not lost.
We will end up changing comment rendering at that point.