link to the wiki even if the doc is not in a shared drive
One may still want to view the doc from the wiki.
Additionally, this could help with inter-document links to docs not in the wiki.

If the user can view it in the docs they can view it in the wiki.
Currently this just takes the first available drive in the manifest.

But we should probably change things so that gdocwiki
recognizes when a document is not in the shared drive.

The view code has been organized into separate view functions.
This allows usage of TypeScript's type narrowing:
the code is now more robust.

The only thing that changed was the link out is now always rendered once loading is complete.