remove unsupported font families
If a font is not supported, a browser will choose a fallback.
But we see bad results from the fallback.

By removing the font-family instead this inherits the font from parents.
If there is no font support in the document itself,
the doc will then use the font of gdocwiki itself.