folder page: user can switch between list and table view
store their switch as a localstorage preference.

Remove the ability to customize the display preference for the folder.
This may be useful in a few cases. However, consider
1) 'hide' was only useful when displaying the README page.
We are now improving this view so that it is unecessary.
2) It can be confusing/frustrating for a user.
This means the user is not in control of how folders are displayed.
Peronsally I always want to see the table view
so for some users this was a bug rather than a feature.