show a spreadsheet as HTML
Since there are still major issues with this,
default to the existing iframe preview.

Add an external link to the preview.
It seems there is no difference between the preview mode
and the HTML export.

So the issue is really just that Google Translate ignores iframe content.
But that is probably not fixable.
It might be better though to fetch the page in question directly
and not have a zip dependency.

TODO:
* vertical scrolling does not work
* support multiple sheets