use a normal link in the table view
For Search, previously files opened in an external tab.
Now that we use a normal link, users can ctrl+click if they want this behavior.

The file name table cell is the link now.
Don't allow clicking other parts of the row.
This will be helpful for clicking on other table elements in the future