fix mobile preview & edit links
on mobile preview & edit don't work
They just diplay the HTML view and they have large paddings.

mobile detection for Google Docs is not based on screen size.
It must be based on either device or touchscreen detection.

Here we do touch screen detection.
On a touch screen, use the Google Docs link.
If the user has the Google app installed this should open in the app.