remove drive selection from the header menu
Just the sidebar for now.
Putting it in the header is nice when there are just a few drives.

However, there can be many drives so that the dropdown would need to scroll.
And other users just have shared folders rather than shared drives.

I think it is better for the header to have the user's favorite locations.